Revert "[Feauture] Added Downloader Object to focus on Downloading files (#20)"
This reverts commit a679941bb5c4ee2a8e9bb635576ed7943dca4058.
