+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     "pandas",
     "bs4",
     "requests",
-    "tqdm",
+    "tqdm"
 ]
 
 classifiers = [
