+++ sec_web_scraper/tests/test_all.py
@@ -1,5 +1,5 @@
 from unittest.mock import patch, MagicMock
-import pandas as pd
+
 import pytest
 import requests
 import os
@@ -82,3 +82,37 @@ def test_create_selenium_browser_headless_con_err():
     with pytest.raises(ConnectionError) as conn_er:
         create_selenium_browser_headless("https://www.sec.gov/edgar/2/2")
         print(f'{conn_er}')
+
+
+# check if directory index_sec exists
+def test_build_sec_pass():
+    # integ test
+    res = build_index_sec(2000, 2002)
+    assert os.path.exists('./index_sec/')
+
+
+@patch('builtins.print')
+def test_build_sec_fail(mock_print):
+    res = build_index_sec(2010, 2011)
+    assert mock_print.call_args.args == ('trying to do Latin encoding',)
+
+
+def test_2011_sec_full_index():
+    # The 2011 Quarter 4 Test should fail
+    sec_url = "https://www.sec.gov/Archives/edgar/full-index/"
+    year = 2011
+    quarter = 4
+    column_names = ['CIK', 'Company Name', 'Form Type', 'Date Filed', 'Filename']
+    dat_types = {"CIK": int, 'Company Name': str, 'Form Type': str, 'Date Filed': str, 'Filename': str}
+    response = requests.get(sec_url + f"{year}/QTR{quarter}/master.zip", headers=headers)
+    assert response.ok
+    with pytest.raises(UnicodeDecodeError) as context:
+        master_index = pd.read_csv(
+            io.BytesIO(response.content),
+            skiprows=11,
+            sep="|",
+            compression='zip',
+            names=column_names,
+            dtype=dat_types,
+        )
+    assert "invalid continuation byte" in str(context.value)
